food_delivery/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
autopep8==2.3.2
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
constantly==23.10.4
coverage==7.13.4
cross-web==0.4.1
cryptography==46.0.5
daphne==4.2.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
Django==6.0.3
django-admin-interface==0.26.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-filter==25.2
django-picklefield==3.4.0
django-q2==1.9.0
django-redis==6.0.0
django-silk==5.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
executing==2.2.1
gprof2dot==2025.4.14
graphql-core==3.2.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kombu==5.6.2
matplotlib-inline==0.2.1
msgpack==1.1.2
nest-asyncio==1.6.0
newrelic==12.0.0
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-django==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.54.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stack-data==0.6.3
strawberry-graphql==0.311.0
strawberry-graphql-django==0.80.0
text-unidecode==1.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wrapt==1.17.3
zipp==3.23.0
zope.interface==8.2
